requirements.txt
ping3
firebase-admin
python-dotenv
